requirements.txt
python-dotenv
supabase
pandas
reflex
python-dotenv

